IscsiClientLib: iscsiadm -m session exit status 21 is "success"
When trying to start iscsid, yast waits for a "iscsiadm -m session"
command to exit with status 0. That fails if there are no sessions
(default situation during installation). The exit status of iscsiadm
in this case is 21: "ISCSI_ERR_NO_OBJS_FOUND - no
records/targets/sessions/portals found to execute operation on".
That should not be treated as an error here. Doing this right
saves 10s useless wait time during installation (actually, 20s,
because iscsid is started again after setting the Initiator name.